+++ src/__tests__/parsers/babel_typescript_parser.test.js
@@ -3,9 +3,13 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ * @flow
  */
 
-import {parseTs} from '../../parsers/babel_parser';
 import {parserTests} from '../../../fixtures/parser_tests';
 
+const {parseOptions} = require('../../parsers/helper');
+const {parse} = require('../../parsers/babel_parser');
+
+const parseTs = (file: string, data?: string) => parse(file, data, parseOptions('always.ts'));
 parserTests(parseTs, true);